pyproject.toml
[project]
name = "gareth-patterson-take-home-challenge"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.11"
dependencies = [
    "annotated-types==0.7.0",
    "anyio==4.9.0",
    "certifi==2025.4.26",
    "charset-normalizer==3.4.2",
    "click==8.1.8",
    "colorama==0.4.6",
    "distro==1.9.0",
    "griffe==1.7.3",
    "h11==0.16.0",
    "httpcore==1.0.9",
    "httpx==0.28.1",
    "httpx-sse==0.4.0",
    "idna==3.10",
    "jiter==0.9.0",
    "mcp[cli]==1.8.0",
    "openai==1.78.0",
    "openai-agents==0.0.14",
    "pinecone[asyncio,grpc]==6.0.2",
    "pinecone-plugin-interface==0.0.7",
    "pydantic==2.11.4",
    "pydantic-core==2.33.2",
    "pydantic-settings==2.9.1",
    "python-dateutil==2.9.0.post0",
    "python-dotenv==1.1.0",
    "python-multipart==0.0.20",
    "requests==2.32.3",
    "six==1.17.0",
    "sniffio==1.3.1",
    "sse-starlette==2.3.4",
    "starlette==0.46.2",
    "types-requests==2.32.0.20250328",
    "typing-extensions==4.13.2",
    "typing-inspection==0.4.0",
    "urllib3==2.4.0",
    "uvicorn==0.34.2",
    "tiktoken>=0.11.0",
    "numpy>=2.3.2",
    "scikit-learn>=1.7.1",
    "diskcache>=5.6.3",
    "plotly>=6.3.0",
    "kaleido>=1.0.0",
    "pyyaml>=6.0.2",
]

[tool.uv.workspace]
members = [
    "gp-mcp-demo",
]

[dependency-groups]
dev = [
    "pytest>=8.4.1",
    "pytest-asyncio>=1.1.0",
]
